reorganize functions, update readme
