+++ readme.md
@@ -1,5 +1,17 @@
 # Notes from Beatrice:
 
+## Re-organized Files
+I reorganized our functions into 4 differente files:
+
+`generateImages.py` generates the desired images using the helper functions contained in `digitalzoom.py` and `dzsynthesis.py`\
+&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;**NOTE: `main.py` should only ever call functions in this file** and not the below helper files:
+
+`digitalzoom.py` contains the digital zoom/reverse digital zoom functions
+
+`dzsynthesis.py` contains both DZ Synthesis functions
+
+`utils.py` contains any utility functions. Right now it just has `show_2_images_side_by_side`
+
 ## Argparser
 I added an arg parser for terminal commands!
